Separated schedule modal to a single widget.
