+++ lib/app.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'app/routes.dart';
 import 'package:provider/provider.dart';
-import 'screens/db_page.dart';
+import 'app/widgets/schedule_modal.dart';
 
 class MyApp extends StatelessWidget {
   const MyApp({super.key});
@@ -57,13 +57,12 @@ class _BottomNavigationState extends State<BottomNavigation> {
         onPressed: () {
           showModalBottomSheet(
             context: context,
-            builder: (context) => const ModalSheetContent(),
+            builder: (context) => const ScheduleModal(),
           );
         },
         child: const Icon(Icons.add),
       )
     : null,
-
       bottomNavigationBar: SafeArea(
         child: BottomAppBar(
           child: Padding(
@@ -196,149 +195,3 @@ class _BottomNavigationState extends State<BottomNavigation> {
   }
 }
 
-class ModalSheetContent extends StatefulWidget {
-  const ModalSheetContent({Key? key}) : super(key: key);
-
-  @override
-  State<ModalSheetContent> createState() => _ModalSheetContentState();
-}
-
-class _ModalSheetContentState extends State<ModalSheetContent> {
-  final _formKey = GlobalKey<FormState>();
-  final _titleController = TextEditingController();
-  final _descriptionController = TextEditingController();
-  DateTime? _selectedDate;
-  TimeOfDay? _selectedTime;
-
-  @override
-  void dispose() {
-    _titleController.dispose();
-    _descriptionController.dispose();
-    super.dispose();
-  }
-
-  Future<void> _pickDate() async {
-    final now = DateTime.now();
-    final DateTime? pickedDate = await showDatePicker(
-      context: context,
-      initialDate: _selectedDate ?? now,
-      firstDate: now,
-      lastDate: DateTime(now.year + 5),
-    );
-    if (pickedDate != null) {
-      setState(() {
-        _selectedDate = pickedDate;
-      });
-    }
-  }
-
-  Future<void> _pickTime() async {
-    final TimeOfDay? pickedTime = await showTimePicker(
-      context: context,
-      initialTime: _selectedTime ?? TimeOfDay.now(),
-    );
-    if (pickedTime != null) {
-      setState(() {
-        _selectedTime = pickedTime;
-      });
-    }
-  }
-
-  void _submitForm() {
-    if (_formKey.currentState!.validate() &&
-        _selectedDate != null &&
-        _selectedTime != null) {
-      final reminderDateTime = DateTime(
-        _selectedDate!.year,
-        _selectedDate!.month,
-        _selectedDate!.day,
-        _selectedTime!.hour,
-        _selectedTime!.minute,
-      );
-      print('Title: ${_titleController.text}');
-      print('Description: ${_descriptionController.text}');
-      print('Reminder Date and Time: $reminderDateTime');
-      Navigator.pop(context);
-    } else {
-      ScaffoldMessenger.of(context).showSnackBar(
-        const SnackBar(content: Text('Please fill in all fields')),
-      );
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      width: MediaQuery.of(context).size.width,
-      padding: const EdgeInsets.all(28.0),
-      child: SingleChildScrollView(
-        child: Form(
-          key: _formKey,
-          child: Column(
-            mainAxisSize: MainAxisSize.min,
-            children: [
-              TextFormField(
-                controller: _titleController,
-                decoration: const InputDecoration(labelText: 'Title'),
-                validator: (value) {
-                  if (value == null || value.trim().isEmpty) {
-                    return 'Please enter a title';
-                  }
-                  return null;
-                },
-              ),
-              const SizedBox(height: 8),
-              TextFormField(
-                controller: _descriptionController,
-                decoration: const InputDecoration(labelText: 'Description'),
-                validator: (value) {
-                  if (value == null || value.trim().isEmpty) {
-                    return 'Please enter a description';
-                  }
-                  return null;
-                },
-              ),
-              const SizedBox(height: 8),
-              Row(
-                children: [
-                  Expanded(
-                    child: Text(
-                      _selectedDate == null
-                          ? 'No date selected'
-                          : 'Date: ${_selectedDate!.toLocal().toString().split(' ')[0]}',
-                    ),
-                  ),
-                  TextButton(
-                    onPressed: _pickDate,
-                    child: const Text('Select Date'),
-                  ),
-                ],
-              ),
-              const SizedBox(height: 8),
-              Row(
-                children: [
-                  Expanded(
-                    child: Text(
-                      _selectedTime == null
-                          ? 'No time selected'
-                          : 'Time: ${_selectedTime!.format(context)}',
-                    ),
-                  ),
-                  TextButton(
-                    onPressed: _pickTime,
-                    child: const Text('Select Time'),
-                  ),
-                ],
-              ),
-              const SizedBox(height: 16),
-              ElevatedButton(
-                onPressed: _submitForm,
-                child: const Text('Save Reminder'),
-              ),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
